Split tasks into separate files
